Move styles from index.js, add fullScreenPortal on mobile, enable clearing dates
